Refactor all fullscreen-related logic into its own hook/file
- follows other toolbar/control behavior
- makes it easier to unit test
- makes it easier to track different full-screen related logic/events in one place
