+++ src/components/datagrid/controls/index.ts
@@ -9,6 +9,7 @@
 export { useDataGridColumnSelector } from './column_selector';
 export { useDataGridColumnSorting } from './column_sorting';
 export { useDataGridDisplaySelector, startingStyles } from './display_selector';
+export { useDataGridFullScreenSelector } from './full_screen_selector';
 export {
   checkOrDefaultToolBarDisplayOptions,
   EuiDataGridToolbar,
